Testing leftover

+++ src/opencl_program.cl
@@ -273,7 +273,7 @@ __kernel void search(__global ulong* result_a, __global uchar* const slab_a, ulo
 static void write_value(__global uchar* slab_a, ulong const index_a, ulong const value_a)
 {
 	const ulong offset_l = index_a * NP_VALUE_SIZE;
-	slab_a[offset_l + 0] = (uint) (value_a >> 0x00);
+	slab_a[offset_l + 0] = (uchar) (value_a >> 0x00);
 	slab_a[offset_l + 1] = (uchar) (value_a >> 0x08);
 	slab_a[offset_l + 2] = (uchar) (value_a >> 0x10);
 	slab_a[offset_l + 3] = (uchar) (value_a >> 0x18);
